% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_fn.R
\name{getPlayerImage}
\alias{getPlayerImage}
\title{Obtain game logs for a specific player for a specific year}
\usage{
getPlayerImage(player)
}
\arguments{
\item{player}{string denoting NBA player}

\item{year}{integer denoting year to get game logs}
}
\value{
string denoting [www.basketball-reference.com] web URL
}
\description{
Obtain game logs for a specific player for a specific year
}
