% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_fn.R
\name{getPlayers}
\alias{getPlayers}
\title{Obtain list of active NBA players for a specified year}
\usage{
getPlayers(year)
}
\arguments{
\item{year}{integer denoting year between [1947] and [2022]}
}
\value{
list of active NBA players
}
\description{
Obtain list of active NBA players for a specified year
}
