% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_fn.R
\name{getStatsTable}
\alias{getStatsTable}
\title{Obtain stats table for a specific player for a specific year}
\usage{
getStatsTable(player, flag = 2)
}
\arguments{
\item{player}{string denoting NBA player}

\item{flag}{integer denoting which table to retrieve (2 = per game stats, 4 = per 36 stats). defaults to 2}
}
\value{
string denoting [www.basketball-reference.com] web URL
}
\description{
Obtain stats table for a specific player for a specific year
}
