% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_fn.R
\name{getURL}
\alias{getURL}
\title{Obtain web URL for given NBA player name}
\usage{
getURL(player)
}
\arguments{
\item{player}{string denoting NBA player}
}
\value{
string denoting [www.basketball-reference.com] web URL
}
\description{
Obtain web URL for given NBA player name
}
